+++ src/Monetization/Application/ConvertRegionPrices/ConvertRegionPrices.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Imdhemy\GooglePlay\Monetization\Application\ConvertRegionPrices;
 
 use Imdhemy\GooglePlay\Domain\Serializer\NormalizerInterface;
-use Imdhemy\GooglePlay\Monetization\Application\MonetizationRequestFactoryInterface;
 use Imdhemy\GooglePlay\Monetization\Domain\ConvertedPrices;
 use Psr\Http\Client\ClientInterface;
 use Throwable;
@@ -20,7 +19,7 @@ final readonly class ConvertRegionPrices
 {
     public function __construct(
         private ClientInterface $client,
-        private MonetizationRequestFactoryInterface $requestFactory,
+        private ConvertRegionPricesRequestFactoryInterface $requestFactory,
         private NormalizerInterface $normalizer,
     ) {
     }
@@ -31,7 +30,7 @@ final readonly class ConvertRegionPrices
     public function execute(ConvertRegionPricesQuery $query): ConvertedPrices
     {
         try {
-            $request = $this->requestFactory->createConvertRegionPricesRequest($query);
+            $request = $this->requestFactory->create($query);
             $response = $this->client->sendRequest($request);
 
             return $this->normalizer->normalize(data: $response, type: ConvertedPrices::class);
