+++ src/Monetization/Application/ConvertRegionPrices/ConvertRegionPricesRequestFactoryInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Imdhemy\GooglePlay\Monetization\Application\ConvertRegionPrices;
+
+use Psr\Http\Message\RequestInterface;
+
+interface ConvertRegionPricesRequestFactoryInterface
+{
+    public function create(ConvertRegionPricesQuery $query): RequestInterface;
+}
